split the injected logic to a standalone hook
