+++ packages/next-common/hooks/connect/useSubstrateAccounts.js
@@ -26,33 +26,6 @@ export function useSubstrateAccounts({
 
   const [accounts, setAccounts] = useState([]);
 
-  const loadInjectedAccounts = useCallback(
-    async (targetWallet) => {
-      const { web3Enable, web3Accounts } = await import(
-        "@polkadot/extension-dapp"
-      );
-
-      try {
-        await web3Enable("subsquare");
-        const injectedAccounts = reject(await web3Accounts(), {
-          type: ChainTypes.ETHEREUM,
-        });
-
-        if (isMounted()) {
-          setAccounts(
-            normalizedSubstrateAccounts(
-              injectedAccounts,
-              targetWallet?.extensionName,
-            ),
-          );
-        }
-      } catch (e) {
-        dispatch(newErrorToast(e.message));
-      }
-    },
-    [isMounted, dispatch],
-  );
-
   const loadPolkadotAccounts = useCallback(
     async (targetWallet) => {
       setAccounts([]);
@@ -117,14 +90,13 @@ export function useSubstrateAccounts({
           break;
         }
         default: {
-          await loadInjectedAccounts(wallet);
           break;
         }
       }
 
       setLoading(false);
     },
-    [loadInjectedAccounts, loadPolkadotAccounts, loadSignetVault],
+    [loadPolkadotAccounts, loadSignetVault],
   );
 
   useEffect(() => {
